+++ bet.py
@@ -1,0 +1,3 @@
+class Bet:
+    def __init__(self, gambler, winner, money, winning_score = -1):
+        self.gambler, self.winner, self.money, self.winning_score = gambler, winner, money, winning_score
