+++ instructions.py
@@ -1,0 +1,14 @@
+from misc import file_read
+
+class Instructions:
+    FORMAT = file_read("instructions.txt")
+
+    def __init__(self):
+        pass
+
+    def __repr__(self):
+        return self.FORMAT
+
+    def interpret(self, arguments):
+        if len(arguments) == 0:
+            print self
