+++ odds.py
@@ -1,0 +1,21 @@
+class Odds:
+    FORMAT = "{player} {odds}"
+    STR_ODDS = ['1:0', '100:1', '50:1', '32:1', '25:1', '20:1', '15:1', '13:1', '11:1', '10:1',
+        '9:1', '8:1', '7:1', '7:1', '6:1', '6:1', '5:1', '5:1', '4:1', '4:1',
+        '4:1', '4:1', '7:2', '7:2', '3:1', '3:1', '3:1', '8:3', '5:2', '7:3',
+        '7:3', '7:3', '2:1', '2:1', '2:1', '2:1', '5:3', '5:3', '5:3', '3:2',
+        '3:2', '3:2', '4:3', '4:3', '5:4', '5:4', '6:5', '6:5', '1:1', '1:1',
+        '1:1', '1:1', '1:1', '5:6', '5:6', '4:5', '4:5', '3:4', '3:4', '2:3',
+        '2:3', '2:3', '3:5', '3:5', '3:5', '1:2', '1:2', '1:2', '1:2', '3:7',
+        '3:7', '3:7', '2:5', '3:8', '1:3', '1:3', '1:3', '2:7', '2:7', '1:4',
+        '1:4', '1:4', '1:4', '1:5', '1:5', '1:6', '1:6', '1:7', '1:7', '1:8',
+        '1:9', '1:10', '1:11', '1:13', '1:15', '1:20', '1:25', '1:32', '1:50', '1:100', '0:1']
+    
+    def __init__(self, player, proportion):
+        self.player, self.proportion = player, proportion
+
+    def __repr__(self):
+        return self.FORMAT.format(
+            player = self.player.name,
+            odds = self.STR_ODDS[int(round(self.proportion * 100))]
+            )
